+++ changelog.md
@@ -1,5 +1,8 @@
-Changelog
-=========
+# Changelog
 
-**Change Log v2.4**
+**Change Log v2.5**  
+[*] Add compatibility with Sharepoint 2010
+[*] Function "progress" added to $SP.list().add()
+
+**Change Log v2.4**  
 [*] First public release